requirements.txt
pandas==2.3.0
pdfplumber==0.11.7
redis==6.2.0
python-dateutil==2.9.0